+++ demo/imports/components/TransitionButton.jsx
@@ -1,4 +1,5 @@
-import React, { PropTypes as pt } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
 import { pure } from 'meteor/ssrwpo:ssr';
 
@@ -6,8 +7,8 @@ const TransitionButton = ({ href, history, children }) => (
   <button onClick={() => history.push(href)}>{children}</button>
 );
 TransitionButton.propTypes = {
-  href: pt.string.isRequired,
-  history: pt.object.isRequired,
-  children: pt.oneOfType([pt.string, pt.element]).isRequired,
+  href: PropTypes.string.isRequired,
+  history: PropTypes.object.isRequired,
+  children: PropTypes.oneOfType([PropTypes.string, PropTypes.element]).isRequired,
 };
 export default withRouter(pure(TransitionButton));