+++ demo/imports/routes/Login.jsx
@@ -1,4 +1,5 @@
-import React, { PureComponent, PropTypes as pt } from 'react';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 import { valueSet } from 'meteor/ssrwpo:ssr';
 import Helmet from 'react-helmet';
 import { connect } from 'react-redux';
@@ -6,9 +7,9 @@ import { Redirect } from 'react-router-dom';
 
 class Login extends PureComponent {
   static propTypes = {
-    isLoggedIn: pt.bool.isRequired,
-    login: pt.func.isRequired,
-    location: pt.object.isRequired,
+    isLoggedIn: PropTypes.bool.isRequired,
+    login: PropTypes.func.isRequired,
+    location: PropTypes.object.isRequired,
   }
   state = { redirectToReferrer: false }
   onLogin = (e) => {